common type refactor
