+++ src/lib/types.ts
@@ -5,10 +5,7 @@ import { DragEvent } from "react";
 import { SelectOption } from "./components/inputs/SelectInput";
 import { View } from "./components/nav/Navigation";
 import { Store } from "./store/types";
-import { DayProps } from "./views/Day";
 import { StateItem } from "./views/Editor";
-import { MonthProps } from "./views/Month";
-import { WeekProps } from "./views/Week";
 import type { RRule } from "rrule";
 import type { JSX } from "react";
 
@@ -38,6 +35,33 @@ export type DayHours =
   | 22
   | 23
   | 24;
+
+export type WeekDays = 0 | 1 | 2 | 3 | 4 | 5 | 6;
+
+interface CommonWeekViewProps {
+  weekDays: WeekDays[];
+  weekStartOn: WeekDays;
+  disableGoToDay?: boolean;
+}
+
+interface CommonViewProps {
+  startHour: DayHours;
+  endHour: DayHours;
+  cellRenderer?(props: CellRenderedProps): JSX.Element;
+  headRenderer?(day: Date): JSX.Element;
+  navigation?: boolean;
+  step: number;
+}
+
+export interface MonthProps extends CommonWeekViewProps, CommonViewProps {}
+
+export interface WeekProps extends CommonWeekViewProps, CommonViewProps {
+  hourRenderer?(hour: string): JSX.Element;
+}
+
+export interface DayProps extends CommonViewProps {
+  hourRenderer?(hour: string): JSX.Element;
+}
 
 export interface CellRenderedProps {
   day: Date;
